move topo utils to analysis
